Consolidate BatchConfiguration changes
