+++ common/src/main/java/org/cbioportal/cmo/pipelines/common/util/CommonBatchConfig.java
@@ -1,0 +1,105 @@
+package org.cbioportal.cmo.pipelines.common.util;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.net.MalformedURLException;
+import javax.sql.DataSource;
+import org.springframework.batch.core.launch.JobLauncher;
+import org.springframework.batch.core.launch.support.SimpleJobLauncher;
+import org.springframework.batch.core.repository.JobRepository;
+import org.springframework.batch.core.repository.dao.Jackson2ExecutionContextStringSerializer;
+import org.springframework.batch.core.repository.dao.XStreamExecutionContextStringSerializer;
+import org.springframework.batch.core.repository.support.JobRepositoryFactoryBean;
+import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.Resource;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.jdbc.datasource.init.DataSourceInitializer;
+import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
+import org.springframework.transaction.PlatformTransactionManager;
+
+/**
+ *
+ * @author ochoaa
+ */
+@Configuration
+public class CommonBatchConfig {
+
+    // general spring batch configuration
+    @Value("org/springframework/batch/core/schema-drop-sqlite.sql")
+    private Resource dropRepositoryTables;
+
+    @Value("org/springframework/batch/core/schema-sqlite.sql")
+    private Resource dataRepositorySchema;
+
+    /**
+     * Spring Batch datasource.
+     * @return DataSource
+     */
+    @Bean
+    public DataSource dataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName("org.sqlite.JDBC");
+        dataSource.setUrl("jdbc:sqlite:repository.sqlite");
+        return dataSource;
+    }
+
+    /**
+     * Spring Batch datasource initializer.
+     * @param dataSource
+     * @return DataSourceInitializer
+     * @throws MalformedURLException
+     */
+    @Bean
+    public DataSourceInitializer dataSourceInitializer(DataSource dataSource) throws MalformedURLException {
+        ResourceDatabasePopulator databasePopulator = new ResourceDatabasePopulator();
+        databasePopulator.addScript(dropRepositoryTables);
+        databasePopulator.addScript(dataRepositorySchema);
+        databasePopulator.setIgnoreFailedDrops(true);
+
+        DataSourceInitializer initializer = new DataSourceInitializer();
+        initializer.setDataSource(dataSource);
+        initializer.setDatabasePopulator(databasePopulator);
+        return initializer;
+    }
+
+    /**
+     * Spring Batch job repository.
+     * @return JobRepository
+     * @throws Exception
+     */
+    private JobRepository getJobRepository() throws Exception {
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.enableDefaultTyping();
+        Jackson2ExecutionContextStringSerializer serializer = new Jackson2ExecutionContextStringSerializer();
+        serializer.setObjectMapper(mapper);
+        
+        JobRepositoryFactoryBean factory = new JobRepositoryFactoryBean();
+        factory.setDataSource(dataSource());
+        factory.setTransactionManager(getTransactionManager());
+        factory.setSerializer(serializer);
+        factory.afterPropertiesSet();
+        return (JobRepository) factory.getObject();
+    }
+
+    /**
+     * Spring Batch transaction manager.
+     * @return PlatformTransactionManager
+     */
+    private PlatformTransactionManager getTransactionManager() {
+        return new ResourcelessTransactionManager();
+    }
+
+    /**
+     * Spring Batch job launcher.
+     * @return JobLauncher
+     * @throws Exception
+     */
+    public JobLauncher getJobLauncher() throws Exception {
+        SimpleJobLauncher jobLauncher = new SimpleJobLauncher();
+        jobLauncher.setJobRepository(getJobRepository());
+        jobLauncher.afterPropertiesSet();
+        return jobLauncher;
+    }
+}
