+++ crdb/src/main/java/org/mskcc/cmo/ks/crdb/pipeline/BatchConfiguration.java
@@ -32,8 +32,6 @@
 
 package org.mskcc.cmo.ks.crdb.pipeline;
 
-import java.net.MalformedURLException;
-import javax.sql.DataSource;
 import org.cbioportal.cmo.pipelines.common.util.EmailUtil;
 import org.mskcc.cmo.ks.crdb.pipeline.model.CRDBDataset;
 import org.mskcc.cmo.ks.crdb.pipeline.model.CRDBPDXClinicalAnnotationMapping;
@@ -45,20 +43,10 @@ import org.mskcc.cmo.ks.crdb.pipeline.model.CRDBSurvey;
 import org.mskcc.cmo.ks.crdb.pipeline.util.CRDBUtils;
 import org.springframework.batch.core.*;
 import org.springframework.batch.core.configuration.annotation.*;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.launch.support.SimpleJobLauncher;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.repository.support.JobRepositoryFactoryBean;
 import org.springframework.batch.item.*;
-import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.*;
-import org.springframework.core.io.Resource;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.jdbc.datasource.init.DataSourceInitializer;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-import org.springframework.transaction.PlatformTransactionManager;
+import org.cbioportal.cmo.pipelines.common.util.CommonBatchConfig;
 
 /**
  * Configuration for running the CRDB clinical data fetcher.
@@ -68,6 +56,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 
 @Configuration
 @EnableBatchProcessing
+@Import(CommonBatchConfig.class)
 public class BatchConfiguration {
 
     public static final String CRDB_IMPACT_JOB = "crdbImpactJob";
@@ -314,76 +303,4 @@ public class BatchConfiguration {
     public StepExecutionListener crdbPDXTimelineListener() {
         return new CRDBPDXTimelineListener();
     }
-
-    // general spring batch configuration
-    @Value("org/springframework/batch/core/schema-drop-sqlite.sql")
-    private Resource dropRepositoryTables;
-
-    @Value("org/springframework/batch/core/schema-sqlite.sql")
-    private Resource dataRepositorySchema;
-
-    /**
-     * Spring Batch datasource.
-     * @return DataSource
-     */
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("org.sqlite.JDBC");
-        dataSource.setUrl("jdbc:sqlite:repository.sqlite");
-        return dataSource;
-    }
-
-    /**
-     * Spring Batch datasource initializer.
-     * @param dataSource
-     * @return DataSourceInitializer
-     * @throws MalformedURLException
-     */
-    @Bean
-    public DataSourceInitializer dataSourceInitializer(DataSource dataSource) throws MalformedURLException {
-        ResourceDatabasePopulator databasePopulator = new ResourceDatabasePopulator();
-        databasePopulator.addScript(dropRepositoryTables);
-        databasePopulator.addScript(dataRepositorySchema);
-        databasePopulator.setIgnoreFailedDrops(true);
-
-        DataSourceInitializer initializer = new DataSourceInitializer();
-        initializer.setDataSource(dataSource);
-        initializer.setDatabasePopulator(databasePopulator);
-        return initializer;
-    }
-
-    /**
-     * Spring Batch job repository.
-     * @return JobRepository
-     * @throws Exception
-     */
-    private JobRepository getJobRepository() throws Exception {
-        JobRepositoryFactoryBean factory = new JobRepositoryFactoryBean();
-        factory.setDataSource(dataSource());
-        factory.setTransactionManager(getTransactionManager());
-        factory.afterPropertiesSet();
-        return (JobRepository) factory.getObject();
-    }
-
-    /**
-     * Spring Batch transaction manager.
-     * @return PlatformTransactionManager
-     */
-    private PlatformTransactionManager getTransactionManager() {
-        return new ResourcelessTransactionManager();
-    }
-
-    /**
-     * Spring Batch job launcher.
-     * @return JobLauncher
-     * @throws Exception
-     */
-    public JobLauncher getJobLauncher() throws Exception {
-        SimpleJobLauncher jobLauncher = new SimpleJobLauncher();
-        jobLauncher.setJobRepository(getJobRepository());
-        jobLauncher.afterPropertiesSet();
-        return jobLauncher;
-    }
 }
-
