+++ darwin/src/main/java/org/mskcc/cmo/ks/darwin/pipeline/BatchConfiguration.java
@@ -32,7 +32,6 @@
 
 package org.mskcc.cmo.ks.darwin.pipeline;
 
-import java.net.MalformedURLException;
 import org.mskcc.cmo.ks.darwin.pipeline.mskimpactdemographics.MskimpactAgeWriter;
 import org.mskcc.cmo.ks.darwin.pipeline.model.*;
 import org.mskcc.cmo.ks.darwin.pipeline.mskimpactbrainspineclinical.*;
@@ -45,27 +44,17 @@ import org.mskcc.cmo.ks.darwin.pipeline.skcm_mskcc_2015_chanttimeline.*;
 import org.mskcc.cmo.ks.darwin.pipeline.util.DarwinUtils;
 
 import java.util.*;
-import javax.sql.DataSource;
+import org.cbioportal.cmo.pipelines.common.util.CommonBatchConfig;
 import org.springframework.batch.core.*;
 import org.springframework.batch.item.*;
 import org.springframework.batch.core.configuration.annotation.*;
 import org.springframework.context.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.batch.core.configuration.annotation.StepScope;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.launch.support.SimpleJobLauncher;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.repository.support.JobRepositoryFactoryBean;
 import org.springframework.batch.core.step.tasklet.Tasklet;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.batch.item.support.*;
-import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.core.io.Resource;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.jdbc.datasource.init.DataSourceInitializer;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-import org.springframework.transaction.PlatformTransactionManager;
 
 /**
  *
@@ -74,6 +63,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 @Configuration
 @EnableBatchProcessing
 @ComponentScan(basePackages="org.mskcc.cmo.ks.redcap")
+@Import(CommonBatchConfig.class)
 public class BatchConfiguration {
 
     public static final String MSKIMPACT_JOB = "mskimpactJob";
@@ -490,76 +480,5 @@ public class BatchConfiguration {
     @StepScope
     public Skcm_mskcc_2015_chantTimelineWriter skcm_mskcc_2015_chantTimelineWriter() {
         return new Skcm_mskcc_2015_chantTimelineWriter();
-    }
-
-    // general spring batch configuration
-    @Value("org/springframework/batch/core/schema-drop-sqlite.sql")
-    private Resource dropRepositoryTables;
-
-    @Value("org/springframework/batch/core/schema-sqlite.sql")
-    private Resource dataRepositorySchema;
-
-    /**
-     * Spring Batch datasource.
-     * @return DataSource
-     */
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("org.sqlite.JDBC");
-        dataSource.setUrl("jdbc:sqlite:repository.sqlite");
-        return dataSource;
-    }
-
-    /**
-     * Spring Batch datasource initializer.
-     * @param dataSource
-     * @return DataSourceInitializer
-     * @throws MalformedURLException
-     */
-    @Bean
-    public DataSourceInitializer dataSourceInitializer(DataSource dataSource) throws MalformedURLException {
-        ResourceDatabasePopulator databasePopulator = new ResourceDatabasePopulator();
-        databasePopulator.addScript(dropRepositoryTables);
-        databasePopulator.addScript(dataRepositorySchema);
-        databasePopulator.setIgnoreFailedDrops(true);
-
-        DataSourceInitializer initializer = new DataSourceInitializer();
-        initializer.setDataSource(dataSource);
-        initializer.setDatabasePopulator(databasePopulator);
-        return initializer;
-    }
-
-    /**
-     * Spring Batch job repository.
-     * @return JobRepository
-     * @throws Exception
-     */
-    private JobRepository getJobRepository() throws Exception {
-        JobRepositoryFactoryBean factory = new JobRepositoryFactoryBean();
-        factory.setDataSource(dataSource());
-        factory.setTransactionManager(getTransactionManager());
-        factory.afterPropertiesSet();
-        return (JobRepository) factory.getObject();
-    }
-
-    /**
-     * Spring Batch transaction manager.
-     * @return PlatformTransactionManager
-     */
-    private PlatformTransactionManager getTransactionManager() {
-        return new ResourcelessTransactionManager();
-    }
-
-    /**
-     * Spring Batch job launcher.
-     * @return JobLauncher
-     * @throws Exception
-     */
-    public JobLauncher getJobLauncher() throws Exception {
-        SimpleJobLauncher jobLauncher = new SimpleJobLauncher();
-        jobLauncher.setJobRepository(getJobRepository());
-        jobLauncher.afterPropertiesSet();
-        return jobLauncher;
     }
 }
