+++ gene/src/main/java/org/mskcc/cmo/ks/gene/config/BatchConfiguration.java
@@ -32,26 +32,15 @@
 
 package org.mskcc.cmo.ks.gene.config;
 
-import java.net.MalformedURLException;
-import javax.sql.DataSource;
 import org.mskcc.cmo.ks.gene.*;
 import org.mskcc.cmo.ks.gene.model.Gene;
 
 import org.springframework.batch.core.*;
 import org.springframework.batch.core.configuration.annotation.*;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.launch.support.SimpleJobLauncher;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.repository.support.JobRepositoryFactoryBean;
 import org.springframework.batch.item.*;
-import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
 import org.springframework.beans.factory.annotation.*;
 import org.springframework.context.annotation.*;
-import org.springframework.core.io.Resource;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.jdbc.datasource.init.DataSourceInitializer;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-import org.springframework.transaction.PlatformTransactionManager;
+import org.cbioportal.cmo.pipelines.common.util.CommonBatchConfig;
 
 /**
  *
@@ -59,6 +48,7 @@ import org.springframework.transaction.PlatformTransactionManager;
  */
 @Configuration
 @EnableBatchProcessing
+@Import(CommonBatchConfig.class)
 public class BatchConfiguration {
 
     public static final String GENE_DATA_JOB = "geneDataJob";
@@ -111,76 +101,5 @@ public class BatchConfiguration {
     @Bean
     public StepExecutionListener geneDataListener() {
         return new GeneDataListener();
-    }
-
-    // general spring batch configuration
-    @Value("org/springframework/batch/core/schema-drop-sqlite.sql")
-    private Resource dropRepositoryTables;
-
-    @Value("org/springframework/batch/core/schema-sqlite.sql")
-    private Resource dataRepositorySchema;
-
-    /**
-     * Spring Batch datasource.
-     * @return DataSource
-     */
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("org.sqlite.JDBC");
-        dataSource.setUrl("jdbc:sqlite:repository.sqlite");
-        return dataSource;
-    }
-
-    /**
-     * Spring Batch datasource initializer.
-     * @param dataSource
-     * @return DataSourceInitializer
-     * @throws MalformedURLException
-     */
-    @Bean
-    public DataSourceInitializer dataSourceInitializer(DataSource dataSource) throws MalformedURLException {
-        ResourceDatabasePopulator databasePopulator = new ResourceDatabasePopulator();
-        databasePopulator.addScript(dropRepositoryTables);
-        databasePopulator.addScript(dataRepositorySchema);
-        databasePopulator.setIgnoreFailedDrops(true);
-
-        DataSourceInitializer initializer = new DataSourceInitializer();
-        initializer.setDataSource(dataSource);
-        initializer.setDatabasePopulator(databasePopulator);
-        return initializer;
-    }
-
-    /**
-     * Spring Batch job repository.
-     * @return JobRepository
-     * @throws Exception
-     */
-    private JobRepository getJobRepository() throws Exception {
-        JobRepositoryFactoryBean factory = new JobRepositoryFactoryBean();
-        factory.setDataSource(dataSource());
-        factory.setTransactionManager(getTransactionManager());
-        factory.afterPropertiesSet();
-        return (JobRepository) factory.getObject();
-    }
-
-    /**
-     * Spring Batch transaction manager.
-     * @return PlatformTransactionManager
-     */
-    private PlatformTransactionManager getTransactionManager() {
-        return new ResourcelessTransactionManager();
-    }
-
-    /**
-     * Spring Batch job launcher.
-     * @return JobLauncher
-     * @throws Exception
-     */
-    public JobLauncher getJobLauncher() throws Exception {
-        SimpleJobLauncher jobLauncher = new SimpleJobLauncher();
-        jobLauncher.setJobRepository(getJobRepository());
-        jobLauncher.afterPropertiesSet();
-        return jobLauncher;
     }
 }
