+++ redcap/redcap_pipeline/pom.xml
@@ -22,6 +22,11 @@
       <version>0.1.0</version>
       <type>jar</type>
     </dependency>
+    <dependency>
+      <groupId>org.mskcc.cmo.ks</groupId>
+      <artifactId>common</artifactId>
+      <version>0.1.0</version>
+    </dependency>
   </dependencies>
 
 </project>
