+++ redcap/redcap_pipeline/src/main/java/org/mskcc/cmo/ks/redcap/pipeline/BatchConfiguration.java
@@ -31,30 +31,19 @@
 */
 package org.mskcc.cmo.ks.redcap.pipeline;
 
-import java.net.MalformedURLException;
 import java.util.*;
-import javax.sql.DataSource;
 import org.apache.log4j.Logger;
+import org.cbioportal.cmo.pipelines.common.util.CommonBatchConfig;
 import org.springframework.batch.core.*;
 import org.springframework.batch.item.*;
 import org.springframework.batch.core.configuration.annotation.*;
 import org.springframework.context.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.batch.core.configuration.annotation.StepScope;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.launch.support.SimpleJobLauncher;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.repository.support.JobRepositoryFactoryBean;
 import org.springframework.batch.core.step.tasklet.Tasklet;
 import org.springframework.batch.item.support.CompositeItemProcessor;
 import org.springframework.batch.item.support.CompositeItemWriter;
-import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.Resource;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.jdbc.datasource.init.DataSourceInitializer;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-import org.springframework.transaction.PlatformTransactionManager;
 
 /**
  *
@@ -65,6 +54,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 @EnableBatchProcessing
 @ComponentScan(basePackages="org.mskcc.cmo.ks.redcap.source.internal")
 @PropertySource("classpath:application.properties")
+@Import(CommonBatchConfig.class)
 public class BatchConfiguration {
     public static final String REDCAP_EXPORT_JOB = "redcapExportJob";
     public static final String REDCAP_RAW_EXPORT_JOB = "redcapRawExportJob";
@@ -286,76 +276,5 @@ public class BatchConfiguration {
     @Bean
     public RawTimelineDataStepListener exportRawTimelineDataStepListener() {
         return new RawTimelineDataStepListener();
-    }
-
-    // general spring batch configuration
-    @Value("org/springframework/batch/core/schema-drop-sqlite.sql")
-    private Resource dropRepositoryTables;
-
-    @Value("org/springframework/batch/core/schema-sqlite.sql")
-    private Resource dataRepositorySchema;
-
-    /**
-     * Spring Batch datasource.
-     * @return DataSource
-     */
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("org.sqlite.JDBC");
-        dataSource.setUrl("jdbc:sqlite:repository.sqlite");
-        return dataSource;
-    }
-
-    /**
-     * Spring Batch datasource initializer.
-     * @param dataSource
-     * @return DataSourceInitializer
-     * @throws MalformedURLException
-     */
-    @Bean
-    public DataSourceInitializer dataSourceInitializer(DataSource dataSource) throws MalformedURLException {
-        ResourceDatabasePopulator databasePopulator = new ResourceDatabasePopulator();
-        databasePopulator.addScript(dropRepositoryTables);
-        databasePopulator.addScript(dataRepositorySchema);
-        databasePopulator.setIgnoreFailedDrops(true);
-
-        DataSourceInitializer initializer = new DataSourceInitializer();
-        initializer.setDataSource(dataSource);
-        initializer.setDatabasePopulator(databasePopulator);
-        return initializer;
-    }
-
-    /**
-     * Spring Batch job repository.
-     * @return JobRepository
-     * @throws Exception
-     */
-    private JobRepository getJobRepository() throws Exception {
-        JobRepositoryFactoryBean factory = new JobRepositoryFactoryBean();
-        factory.setDataSource(dataSource());
-        factory.setTransactionManager(getTransactionManager());
-        factory.afterPropertiesSet();
-        return (JobRepository) factory.getObject();
-    }
-
-    /**
-     * Spring Batch transaction manager.
-     * @return PlatformTransactionManager
-     */
-    private PlatformTransactionManager getTransactionManager() {
-        return new ResourcelessTransactionManager();
-    }
-
-    /**
-     * Spring Batch job launcher.
-     * @return JobLauncher
-     * @throws Exception
-     */
-    public JobLauncher getJobLauncher() throws Exception {
-        SimpleJobLauncher jobLauncher = new SimpleJobLauncher();
-        jobLauncher.setJobRepository(getJobRepository());
-        jobLauncher.afterPropertiesSet();
-        return jobLauncher;
     }
 }
